+++ tools/addstring.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
-grep -RIl "\<string name=\"msg_service_updating" res | xargs sed -i -e '/msg_service_updating/d'
+grep -RIl "\<string name=\"help_settings" res | xargs sed -i -e '/help_settings/d'
 
-grep -RIl "\<string name=\"msg_service_version" res | xargs sed -i -e '/msg_service_version/a \
-\ \ \ \ <string name=\"msg_service_updating\">Please wait until the update service has been completed</string>'
+grep -RIl "\<string name=\"help_shared" res | xargs sed -i -e '/help_shared/a \
+\ \ \ \ <string name=\"help_settings\">has application specific fake values</string>'
 
 #grep -RIl "\<string name=\"app_notxposed" res | xargs sed -i -e '/app_notxposed/d'
 #grep -RIl "\<string name=\"restrict_help_internet" res | xargs sed -i -e 's/internet</Internet</g'